requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
distro==1.9.0
docstring_parser==0.16
fastapi==0.112.0
filelock==3.15.4
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instructor==1.3.7
jiter==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.38.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tenacity==8.5.0
tldextract==5.1.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
websockets==12.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
